+++ packages/components/index.ts
@@ -1,0 +1,3 @@
+import XButton from "./button";
+
+export { XButton };