+++ src/App.vue
@@ -12,6 +12,7 @@ import HelloWorld from "./components/HelloWorld.vue";
     </a>
   </div>
   <HelloWorld msg="Vite + Vue" />
+  <x-button theme="info"></x-button>
 </template>
 
 <style scoped>
@@ -21,9 +22,11 @@ import HelloWorld from "./components/HelloWorld.vue";
   will-change: filter;
   transition: filter 300ms;
 }
+
 .logo:hover {
   filter: drop-shadow(0 0 2em #646cffaa);
 }
+
 .logo.vue:hover {
   filter: drop-shadow(0 0 2em #42b883aa);
 }