+++ src/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from "vue";
 import "./style.css";
 import App from "./App.vue";
+import chenxing from "../index";
 
-createApp(App).mount("#app");
+createApp(App).use(chenxing).mount("#app");