+++ vite.config.ts
@@ -1,9 +1,10 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import path from "path";
+import vueJsxPlugin from "@vitejs/plugin-vue-jsx";
 
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [vue(), vueJsxPlugin()],
   resolve: {
     // 配置路径别名
     alias: {